Fix touchpad scrolling

+++ src/main.cpp
@@ -272,8 +272,8 @@ LRESULT CALLBACK WndProc(HWND hwnd, UINT msg, WPARAM wparam, LPARAM lparam) {
 		};
 		ScreenToClient(hwnd, &client_point);
 
-		short wheel_distance = GET_WHEEL_DELTA_WPARAM(wparam);
-		short scroll_amount = wheel_distance / WHEEL_DELTA;
+		float wheel_distance = GET_WHEEL_DELTA_WPARAM(wparam);
+		float scroll_amount = wheel_distance / WHEEL_DELTA;
 		auto [row, col] = RendererCursorToGridPoint(context->renderer, client_point.x, client_point.y);
 		MouseAction action = scroll_amount > 0 ? MouseAction::MouseWheelUp : MouseAction::MouseWheelDown;
